requirements.txt
ccxt~=1.29.15
docopt~=0.6.2
python-dateutil~=2.8.1
plotly~=4.8.1
pandas~=1.0.4
setuptools~=47.1.1
numpy~=1.18.4
requests~=2.23.0
python-dotenv~=0.13.0
websocket-client~=0.57.0
pyTelegramBotAPI~=3.7.1
coverage~=5.1
websockets~=8.1